Add explicit casts, replace constant L suffix by U, fix a few variable hiding
git-svn-id: https://svn.osgeo.org/gdal/trunk@32234 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/alg/gdalwarper.cpp
@@ -815,7 +815,7 @@ GDALWarpDstAlphaMasker( void *pMaskFuncArg, int nBandCount,
         eErr = GDALRasterIO( hAlphaBand, GF_Write, 
                              nXOff, nYOff, nDstXSize, nDstYSize, 
                              pafMask, nDstXSize, nDstYSize, GDT_Float32,
-                             0, sizeof(float) * nXSize );
+                             0, (int)sizeof(float) * nXSize );
         return eErr;
     }
 }
